// js/stat-evaluation-player/src/generated/BallCarryStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LabeledCounts } from "./LabeledCounts.ts";

/**
 * Accumulated ball-carry stats: counts, times, and distances.
 */
export type BallCarryStats = { carry_count: number, total_carry_time: number, total_straight_line_distance: number, total_path_distance: number, longest_carry_time: number, furthest_carry_distance: number, fastest_carry_speed: number, carry_speed_sum: number, average_horizontal_gap_sum: number, average_vertical_gap_sum: number, labeled_event_counts?: LabeledCounts, };